Card styles and code refactor.
