--- src/plays/news-feed-application/data.js
@@ -1,661 +1,0 @@
-export const categories = ["business", "health", "entertainment", "science", "sports", "technology", "general"];
-
-export const countries = [
-    {
-      code: "AR",
-      code3: "ARG",
-      name: "Argentina",
-      number: "032",
-    },
-    {
-      code: "AU",
-      code3: "AUS",
-      name: "Australia",
-      number: "036",
-    },
-    {
-      code: "AT",
-      code3: "AUT",
-      name: "Austria",
-      number: "040",
-    },
-    {
-      code: "BE",
-      code3: "BEL",
-      name: "Belgium",
-      number: "056",
-    },
-    {
-      code: "BR",
-      code3: "BRA",
-      name: "Brazil",
-      number: "076",
-    },
-    {
-      code: "BG",
-      code3: "BGR",
-      name: "Bulgaria",
-      number: "100",
-    },
-    {
-      code: "CA",
-      code3: "CAN",
-      name: "Canada",
-      number: "124",
-    },
-    {
-      code: "CN",
-      code3: "CHN",
-      name: "China",
-      number: "156",
-    },
-    {
-      code: "CO",
-      code3: "COL",
-      name: "Colombia",
-      number: "170",
-    },
-    {
-      code: "CU",
-      code3: "CUB",
-      name: "Cuba",
-      number: "192",
-    },
-    {
-      code: "CZ",
-      code3: "CZE",
-      name: "Czechia",
-      number: "203",
-    },
-    {
-      code: "EG",
-      code3: "EGY",
-      name: "Egypt",
-      number: "818",
-    },
-    {
-      code: "FR",
-      code3: "FRA",
-      name: "France",
-      number: "250",
-    },
-    {
-      code: "DE",
-      code3: "DEU",
-      name: "Germany",
-      number: "276",
-    },
-    {
-      code: "GR",
-      code3: "GRC",
-      name: "Greece",
-      number: "300",
-    },
-    {
-      code: "HK",
-      code3: "HKG",
-      name: "Hong Kong",
-      number: "344",
-    },
-    {
-      code: "HU",
-      code3: "HUN",
-      name: "Hungary",
-      number: "348",
-    },
-    {
-      code: "IN",
-      code3: "IND",
-      name: "India",
-      number: "356",
-    },
-    {
-      code: "ID",
-      code3: "IDN",
-      name: "Indonesia",
-      number: "360",
-    },
-    {
-      code: "IE",
-      code3: "IRL",
-      name: "Ireland",
-      number: "372",
-    },
-    {
-      code: "IL",
-      code3: "ISR",
-      name: "Israel",
-      number: "376",
-    },
-    {
-      code: "IT",
-      code3: "ITA",
-      name: "Italy",
-      number: "380",
-    },
-    {
-      code: "JP",
-      code3: "JPN",
-      name: "Japan",
-      number: "392",
-    },
-    {
-      code: "KR",
-      code3: "KOR",
-      name: "Korea (the Republic of)",
-      number: "410",
-    },
-    {
-      code: "LV",
-      code3: "LVA",
-      name: "Latvia",
-      number: "428",
-    },
-    {
-      code: "LT",
-      code3: "LTU",
-      name: "Lithuania",
-      number: "440",
-    },
-    {
-      code: "MY",
-      code3: "MYS",
-      name: "Malaysia",
-      number: "458",
-    },
-    {
-      code: "MX",
-      code3: "MEX",
-      name: "Mexico",
-      number: "484",
-    },
-    {
-      code: "MA",
-      code3: "MAR",
-      name: "Morocco",
-      number: "504",
-    },
-    {
-      code: "NL",
-      code3: "NLD",
-      name: "Netherlands (the)",
-      number: "528",
-    },
-    {
-      code: "NZ",
-      code3: "NZL",
-      name: "New Zealand",
-      number: "554",
-    },
-    {
-      code: "NG",
-      code3: "NGA",
-      name: "Nigeria",
-      number: "566",
-    },
-    {
-      code: "NO",
-      code3: "NOR",
-      name: "Norway",
-      number: "578",
-    },
-    {
-      code: "PH",
-      code3: "PHL",
-      name: "Philippines (the)",
-      number: "608",
-    },
-    {
-      code: "PL",
-      code3: "POL",
-      name: "Poland",
-      number: "616",
-    },
-    {
-      code: "PT",
-      code3: "PRT",
-      name: "Portugal",
-      number: "620",
-    },
-    {
-      code: "RO",
-      code3: "ROU",
-      name: "Romania",
-      number: "642",
-    },
-    {
-      code: "RU",
-      code3: "RUS",
-      name: "Russian Federation (the)",
-      number: "643",
-    },
-    {
-      code: "SA",
-      code3: "SAU",
-      name: "Saudi Arabia",
-      number: "682",
-    },
-    {
-      code: "RS",
-      code3: "SRB",
-      name: "Serbia",
-      number: "688",
-    },
-    {
-      code: "SG",
-      code3: "SGP",
-      name: "Singapore",
-      number: "702",
-    },
-    {
-      code: "SK",
-      code3: "SVK",
-      name: "Slovakia",
-      number: "703",
-    },
-    {
-      code: "SI",
-      code3: "SVN",
-      name: "Slovenia",
-      number: "705",
-    },
-    {
-      code: "ZA",
-      code3: "ZAF",
-      name: "South Africa",
-      number: "710",
-    },
-    {
-      code: "SE",
-      code3: "SWE",
-      name: "Sweden",
-      number: "752",
-    },
-    {
-      code: "CH",
-      code3: "CHE",
-      name: "Switzerland",
-      number: "756",
-    },
-    {
-      code: "TW",
-      code3: "TWN",
-      name: "Taiwan",
-      number: "158",
-    },
-    {
-      code: "TH",
-      code3: "THA",
-      name: "Thailand",
-      number: "764",
-    },
-    {
-      code: "TR",
-      code3: "TUR",
-      name: "Turkey",
-      number: "792",
-    },
-    {
-      code: "UA",
-      code3: "UKR",
-      name: "Ukraine",
-      number: "804",
-    },
-    {
-      code: "AE",
-      code3: "ARE",
-      name: "United Arab Emirates (the)",
-      number: "784",
-    },
-    {
-      code: "GB",
-      code3: "GBR",
-      name: "United Kingdom of Great Britain and Northern Ireland (the)",
-      number: "826",
-    },
-    {
-      code: "US",
-      code3: "USA",
-      name: "United States of America (the)",
-      number: "840",
-    },
-    {
-      code: "VE",
-      code3: "VEN",
-      name: "Venezuela (Bolivarian Republic of)",
-      number: "862",
-    },
-  ];
-  
-  export const data = [
-    {
-      source: {
-        id: null,
-        name: "ESPN",
-      },
-      author: "Joon Lee",
-      title: "Padres eliminate Mets, secure NLDS date with rival Dodgers - ESPN",
-      description:
-        "Joe Musgrove baffled the New York Mets in Game 3 on Sunday night at Citi Field, leading the San Diego Padres to a 6-0 victory, sealing a berth in the NLDS against the Los Angeles Dodgers.",
-      url: "https://www.espn.com/mlb/story/_/id/34765150/san-diego-padres-eliminate-new-york-mets-secure-nlds-date-los-angeles-dodgers",
-      urlToImage:
-        "https://a.espncdn.com/combiner/i?img=%2Fphoto%2F2022%2F1010%2Fr1073551_1296x729_16%2D9.jpg",
-      publishedAt: "2022-10-10T04:18:45Z",
-      content:
-        "NEW YORK -- San Diego Padres starter Joe Musgrove knew Sunday was the biggest start of his life and felt sick to his stomach all day.\r\nThe stakes were high. A winner-takes-all playoff finale pitching… [+6177 chars]",
-    },
-    {
-      source: {
-        id: "abc-news",
-        name: "ABC News",
-      },
-      author: "MICHELLE L. PRICE Associated Press",
-      title:
-        "NY Rep. Lee Zeldin says 2 people shot in front of his home - ABC News",
-      description:
-        "New York congressman and Republican candidate for governor Lee Zeldin says his family is safe after two teenagers were shot outside his Long Island home",
-      url: "https://abcnews.go.com/Politics/wireStory/ny-rep-lee-zeldin-people-shot-front-home-91260798",
-      urlToImage:
-        "https://s.abcnews.com/images/Politics/WireAP_4cf69e3de0f048b29111a96c220b51ff_16x9_992.jpg",
-      publishedAt: "2022-10-10T03:33:45Z",
-      content:
-        "NEW YORK -- New York congressman and Republican candidate for governor Lee Zeldin says his family is safe after two teenagers were shot outside his Long Island home Sunday afternoon.\r\nThe boys, both … [+2465 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "YouTube",
-      },
-      author: null,
-      title:
-        "Cincinatti Bengals vs. Baltimore Ravens | 2022 Week 5 Highlights - NFL",
-      description:
-        "Check out our other channels:NFL Mundo https://www.youtube.com/mundonflNFL Brasil https://www.youtube.com/c/NFLBrasilOficialNFL UK https://www.youtube.com/ch...",
-      url: "https://www.youtube.com/watch?v=CPM3yJWWYt0",
-      urlToImage: "https://i.ytimg.com/vi/CPM3yJWWYt0/maxresdefault.jpg",
-      publishedAt: "2022-10-10T03:28:48Z",
-      content: null,
-    },
-    {
-      source: {
-        id: "reuters",
-        name: "Reuters",
-      },
-      author: null,
-      title:
-        "Chinese tech shares tumble after U.S. publishes new export rules - Reuters",
-      description:
-        'Shares in Chinese tech giants Alibaba Group <a href="https://www.reuters.com/companies/9988.HK" target="_blank">(9988.HK)</a> and Tencent <a href="https://www.reuters.com/companies/0700.HK" target="_blank">(0700.HK)</a> as well as in chipmakers slumped on Mon…',
-      url: "https://www.reuters.com/technology/chinese-chip-makers-shares-slump-after-us-publishes-new-export-rules-2022-10-10/",
-      urlToImage:
-        "https://www.reuters.com/resizer/WLaI8MTX9zyfa6ubN-HC_bhbf74=/1200x628/smart/filters:quality(80)/cloudfront-us-east-2.images.arcpublishing.com/reuters/XS7K7EAZ2BMG7MTFOU6LP6EEQI.jpg",
-      publishedAt: "2022-10-10T03:14:00Z",
-      content:
-        "SHANGHAI, Oct 10 (Reuters) - Shares in Chinese tech giants Alibaba Group (9988.HK) and Tencent (0700.HK) as well as in chipmakers slumped on Monday, following the latest U.S. crackdown on China's chi… [+1951 chars]",
-    },
-    {
-      source: {
-        id: "cnn",
-        name: "CNN",
-      },
-      author: "Darya Tarasova, Amy Woodyatt",
-      title:
-        "Putin to chair Russia Security Council meeting after humiliating explosion on strategic Crimea bridge - CNN",
-      description:
-        "Russian President Vladimir Putin will hold an operational meeting of his Security Council on Monday, just two days after a massive explosion on a key strategic bridge linking Crimea and Russia.",
-      url: "https://www.cnn.com/2022/10/09/europe/putin-russia-ukraine-security-council-intl/index.html",
-      urlToImage:
-        "https://media.cnn.com/api/v1/images/stellar/prod/221009111200-16-kerch-bridge-crimea-1008.jpg?c=16x9&q=w_800,c_fill",
-      publishedAt: "2022-10-10T02:58:00Z",
-      content:
-        "Russian President Vladimir Putin will hold an operational meeting of his Security Council on Monday, just two days after a massive explosion on a key strategic bridge linking Crimea and Russia.\r\nThe … [+4667 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "YouTube",
-      },
-      author: null,
-      title:
-        "Secret recording controversy: LA City Council President Nury Martinez under criticism for racist rem - CBS Los Angeles",
-      description:
-        "The secret recording was published by the Los Angeles Times on Sunday. It exposes Los Angeles City Council President Nury Martinez using racial language talk...",
-      url: "https://www.youtube.com/watch?v=f6Dzl0qOetA",
-      urlToImage: "https://i.ytimg.com/vi/f6Dzl0qOetA/maxresdefault.jpg",
-      publishedAt: "2022-10-10T02:21:51Z",
-      content: null,
-    },
-    {
-      source: {
-        id: null,
-        name: "TVLine",
-      },
-      author: "Kimberly Roots",
-      title:
-        "‘House of the Dragon’ Recap: Season 1, Episode 8 — [Spoiler] Dies - TVLine",
-      description:
-        "For a brief moment in this week’s House of the Dragon, it looks like King Viserys’ dream of family unity might actually come to pass. Because, really, who’d be able to refuse the last wishes of a visibly dying man who’s merely asking his kin not to destroy ea…",
-      url: "https://tvline.com/2022/10/09/house-of-the-dragon-recap-season-1-episode-8-viserys-dies/",
-      urlToImage:
-        "https://tvline.com/wp-content/uploads/2022/10/house-of-the-dragon-recap-season-1-episode-8-viserys-dies.jpg?w=620",
-      publishedAt: "2022-10-10T02:08:00Z",
-      content:
-        "For a brief moment in this week’s House of the Dragon, it looks like King Viserys’ dream of family unity might actually come to pass. Because, really, who’d be able to refuse the last wishes of a vis… [+11030 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "WESH Orlando",
-      },
-      author: null,
-      title: "Hurricane Julia weakens to tropical storm - WESH 2 Orlando",
-      description: "Hurricane Julia weakened to a tropical storm on Sunday.",
-      url: "https://www.wesh.com/article/hurricane-weakens-tropical-storm-julia/41567845",
-      urlToImage:
-        "https://kubrick.htvapps.com/htv-prod-media.s3.amazonaws.com/images/trackingthetropics-1665364292.jfif?crop=1.00xw:1.00xh;0,0&resize=1200:*",
-      publishedAt: "2022-10-10T01:11:00Z",
-      content:
-        "Hurricane Julia weakened to a tropical storm on Sunday. \r\nAccording to the National Hurricane Center, portions of Central America could experience flash flooding and mudslides on Sunday night and Mon… [+3851 chars]",
-    },
-    {
-      source: {
-        id: "the-washington-post",
-        name: "The Washington Post",
-      },
-      author: "Amy B Wang, Azi Paybarah",
-      title:
-        "Herschel Walker allegations keep abortion in spotlight for midterms - The Washington Post",
-      description:
-        "Abortion has stayed in the headlines with reports that Georgia GOP Senate candidate Herschel Walker paid for an ex-girlfriend’s procedure and later urged her to have a second one.",
-      url: "https://www.washingtonpost.com/politics/2022/10/09/abortion-herschel-walker-midterms/",
-      urlToImage:
-        "https://www.washingtonpost.com/wp-apps/imrs.php?src=https://d1i4t8bqe7zgj6.cloudfront.net/10-09-2022/t_889025c0864d4ecfa999e962662a6ea5_name_20221010_SundayShowSPLIT_STILL.jpg&w=1440",
-      publishedAt: "2022-10-10T00:41:00Z",
-      content:
-        "Bombshell allegations about Herschel Walker, the Republican nominee for U.S. Senate in Georgia, have drawn renewed attention to the debate over abortion, which Democrats are hoping will boost turnout… [+7240 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "CBS Sports",
-      },
-      author: "",
-      title:
-        "NFL Week 5 grades: Giants get an 'A+' for upset in London; Steelers get an 'F' for blowout loss to Bills - CBS Sports",
-      description:
-        "Here are the Week 5 grades for every team that played on Sunday",
-      url: "https://www.cbssports.com/nfl/news/nfl-week-5-grades-giants-get-an-a-for-upset-in-london-steelers-get-an-f-for-blowout-loss-to-bills/",
-      urlToImage:
-        "https://sportshub.cbsistatic.com/i/r/2022/10/09/889f2bde-d92b-4b04-971e-fe164a4fd5eb/thumbnail/1200x675/6a82b3d5265a6ad75c97a3ba7e78e38b/gettyimages-1431988826.jpg",
-      publishedAt: "2022-10-10T00:18:00Z",
-      content:
-        "The Giants have been to London three times and all three times, they've come away with a victory and it's probably safe to say their most recent win was also their most improbable one. \r\nOn paper, it… [+4821 chars]",
-    },
-    {
-      source: {
-        id: "cnn",
-        name: "CNN",
-      },
-      author: "Marlon Sorto, Sahar Akbarzai, Michelle Velez",
-      title:
-        "A landslide in Venezuela has killed at least 22 people and 50 others are missing - CNN",
-      description:
-        "A landslide in Venezuela on Sunday has killed at least 22 people and left more than 50 missing in the north central state of Aragua, according to officials.",
-      url: "https://www.cnn.com/2022/10/09/americas/venezuela-landslide-aragua/index.html",
-      urlToImage:
-        "https://media.cnn.com/api/v1/images/stellar/prod/221009161802-01-venezuela-landslide-1009.jpg?c=16x9&q=w_800,c_fill",
-      publishedAt: "2022-10-09T23:48:00Z",
-      content:
-        "A landslide in Venezuela on Sunday has killed at least 22 people and left more than 50 missing in the north central state of Aragua, according to officials.\r\nThe landslide came down in the Santos Mic… [+700 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "The Guardian",
-      },
-      author: "Julia Kollewe",
-      title:
-        "Cutting oil output risks global economy, warns US Treasury secretary - The Guardian",
-      description:
-        "Janet Yellen’s comments come as figures show business activity declining across most UK regions",
-      url: "https://www.theguardian.com/business/2022/oct/10/cutting-oil-output-risks-global-economy-warns-us-treasury-secretary",
-      urlToImage:
-        "https://i.guim.co.uk/img/media/c8e22ebf4b3435fed19b478a2b552103ba17f85e/0_49_3253_1951/master/3253.jpg?width=1200&height=630&quality=85&auto=format&fit=crop&overlay-align=bottom%2Cleft&overlay-width=100p&overlay-base64=L2ltZy9zdGF0aWMvb3ZlcmxheXMvdGctZGVmYXVsdC5wbmc&enable=upscale&s=da2f176954721bc8f46fb9d1d28eb1c5",
-      publishedAt: "2022-10-09T23:01:00Z",
-      content:
-        "The worlds biggest oil-producing nations cutting production at a time of soaring energy costs is unhelpful and unwise for global economic growth, the US Treasury secretary has warned, amid intense pr… [+4933 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "ESPN",
-      },
-      author: "ESPN staff",
-      title:
-        "Bills, Chargers, Giants among Week 5's top NFL Twitter trolls - ESPN",
-      description: "These teams had their jokes ready after Week 5's action.",
-      url: "https://www.espn.com/nfl/story/_/id/34763573/bills-chargers-giants-week-5-top-nfl-twitter-trolls",
-      urlToImage:
-        "https://a3.espncdn.com/combiner/i?img=%2Fphoto%2F2022%2F1009%2Fr1073448_1022x575_16%2D9.png",
-      publishedAt: "2022-10-09T22:59:45Z",
-      content:
-        "From historic blowouts to surprising upsets and more, Week 5 of the 2022 NFL season has been one for the books.\r\nThe first took place In London as the New York Giants and Green Bay Packers faced off … [+2266 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "Eonline.com",
-      },
-      author: "Corinne Heller",
-      title:
-        "Watch Chicago West and Psalm West Sing Dad Kanye West's Song That References Them - E! NEWS",
-      description:
-        "Kim Kardashian shared a video of her and ex Kanye West's youngest children, Chicago and Psalm, singing his and XXXTentacion's song \"True Love.\"",
-      url: "https://www.eonline.com/news/1349640/watch-chicago-west-and-psalm-west-sing-dad-kanye-wests-song-that-references-them",
-      urlToImage:
-        "https://akns-images.eonline.com/eol_images/Entire_Site/202299/rs_1200x1200-221009132435-1200-chicago-west-instagram-cjh-100922.jpg?fit=around%7C1080:1080&output-quality=90&crop=1080:1080;center,top",
-      publishedAt: "2022-10-09T22:42:00Z",
-      content:
-        "Following in dad's footsteps?\r\nIn an Instagram video shared by Kim Kardashian Oct. 9, her and ex Kanye West's youngest children Chicago West, 4, and Psalm West, 3, perform an adorable rendition of hi… [+974 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "Thetechoutlook.com",
-      },
-      author: "Urvi Gupte",
-      title:
-        "Planning to buy Apple's iPhone 14? : Here's why you should buy it - The Tech Outlook",
-      description:
-        "Apple is always in the news with its latest iPhones. This time, it is the new iPhone 14 by Apple. iPhone 14 is being preferred by many and the purchase has soared high. There are actual reasons why you should buy the iPhone 14 by Apple in the year 20",
-      url: "https://www.thetechoutlook.com/e-commerce/planning-to-buy-apples-iphone-14-heres-why-you-should-buy-it/",
-      urlToImage:
-        "https://www.thetechoutlook.com/wp-content/uploads/2022/10/Untitled-design-73.jpg",
-      publishedAt: "2022-10-09T22:18:36Z",
-      content:
-        "Apple is always in the news with its latest iPhones. This time, it is the new iPhone 14 by Apple. iPhone 14 is being preferred by many and the purchase has soared high. There are actual reasons why y… [+1289 chars]",
-    },
-    {
-      source: {
-        id: "cnn",
-        name: "CNN",
-      },
-      author: "Brenda Goodman",
-      title:
-        "Colonoscopies may not reduce cancer deaths, study finds -- but experts say you should still get one - CNN",
-      description:
-        "Colonoscopy may reduce the risk of dying of colorectal cancer by as much as 50%, but there's a catch: It only works if you get the scan.",
-      url: "https://www.cnn.com/2022/10/09/health/colonoscopy-cancer-death-study/index.html",
-      urlToImage:
-        "https://media.cnn.com/api/v1/images/stellar/prod/221007192334-endoscope-stock.jpg?c=16x9&q=w_800,c_fill",
-      publishedAt: "2022-10-09T22:08:00Z",
-      content:
-        "Colonoscopy may reduce the risk of dying of colorectal cancer by as much as 50%, but theres a catch: It only works if you get the scan.\r\nThats the big takeaway message from the first randomized trial… [+9345 chars]",
-    },
-    {
-      source: {
-        id: "bloomberg",
-        name: "Bloomberg",
-      },
-      author: null,
-      title: "Stock Market Today: Dow, S&P Live Updates for Oct. 10 - Bloomberg",
-      description: null,
-      url: "https://www.bloomberg.com/tosv2.html?vid=&uuid=f6b85053-4857-11ed-93cd-54434f776c6c&url=L25ld3MvYXJ0aWNsZXMvMjAyMi0xMC0wOS9hc2lhbi1zdG9ja3Mtc2V0LXRvLWRlY2xpbmUtYXMtcmF0ZS1oaWtlLWJldHMtZmlybS1tYXJrZXRzLXdyYXA=",
-      urlToImage: null,
-      publishedAt: "2022-10-09T21:39:42Z",
-      content:
-        "To continue, please click the box below to let us know you're not a robot.",
-    },
-    {
-      source: {
-        id: "usa-today",
-        name: "USA Today",
-      },
-      author: "David Rising, USA TODAY",
-      title:
-        "CNN crew fined in Thailand after 'unethical' day care attack coverage - USA TODAY",
-      description:
-        "CNN has pulled a story on the massacre of Thai preschoolers and apologized over criticism its journalists entered the day care without permission.",
-      url: "https://www.usatoday.com/story/news/world/2022/10/09/cnn-crew-fined-thailand-day-care-attack/8230701001/",
-      urlToImage:
-        "https://www.gannett-cdn.com/presto/2022/10/08/USAT/d8440659-df99-4ca1-a2d1-66702f1656b3-AP_APTOPIX_Thailand_Childcare_Center_Shooting.jpg?auto=webp&crop=5999,3375,x0,y305&format=pjpg&width=1200",
-      publishedAt: "2022-10-09T21:31:04Z",
-      content:
-        "UTHAI SAWAN, Thailand — CNN pulled a story on the massacre of Thai preschoolers and apologized Sunday over criticism its journalists entered the day care where the children were slain and filmed the … [+3167 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "The Guardian",
-      },
-      author: "Ramon Antonio Vargas",
-      title:
-        "Republican Chuck Grassley vows to vote against a national abortion ban - The Guardian US",
-      description:
-        "The longest-tenured US senator joins a growing chorus of conservative lawmakers opposed to such a restriction",
-      url: "https://www.theguardian.com/us-news/2022/oct/09/chuck-grassley-republican-against-national-abortion-ban",
-      urlToImage:
-        "https://i.guim.co.uk/img/media/ec86783870f58432c261426e1a2e075520fda0b7/0_182_5488_3293/master/5488.jpg?width=1200&height=630&quality=85&auto=format&fit=crop&overlay-align=bottom%2Cleft&overlay-width=100p&overlay-base64=L2ltZy9zdGF0aWMvb3ZlcmxheXMvdGctZGVmYXVsdC5wbmc&enable=upscale&s=18d9683825b53d71b24440ab1a323649",
-      publishedAt: "2022-10-09T19:18:00Z",
-      content:
-        "The longest-tenured Republican in the US Senate has pledged to vote against a national ban on abortions after 15 weeks of pregnancy which a prominent fellow party member and chamber colleague propose… [+3134 chars]",
-    },
-    {
-      source: {
-        id: null,
-        name: "CNBC",
-      },
-      author: "Ashley Capoot",
-      title:
-        "Allianz Chief Economic Adviser El-Erian believes core inflation 'is still going up' - CNBC",
-      description:
-        'Allianz Chief Economic Adviser Mohamed El-Erian said he predicts headline inflation "will probably come down to about 8%," but that core inflation "is still going up."',
-      url: "https://www.cnbc.com/2022/10/09/allianz-chief-economic-adviser-el-erian-believes-core-inflation-is-still-going-up.html",
-      urlToImage:
-        "https://image.cnbcfm.com/api/v1/image/106216483-1572545732717nup_187507_0327.jpg?v=1665339930&w=1920&h=1080",
-      publishedAt: "2022-10-09T18:25:30Z",
-      content:
-        'Ahead of the release of the latest consumer price index reading this week, Allianz Chief Economic Adviser Mohamed El-Erian told CBS\' "Face The Nation" Sunday that he predicts headline inflation "will… [+878 chars]',
-    },
-  ];
-
