+++ src/plays/news-feed-application/styles.css
@@ -2,8 +2,7 @@
 .card-container {
     display: grid;
     grid-template-columns: 200px 200px 200px 200px 200px;
-    /* grid-auto-flow: row; */
-    grid-gap: 16px;
+    grid-gap: 24px;
     padding: 84px 16px;
     justify-content: space-evenly;
     width: 100%;
@@ -11,6 +10,13 @@
 
 .card {
     cursor: pointer;
+    background-color: #f1f2f3 !important;
+}
+
+.card:hover {
+    transition: transform .2s;
+    transform: scale(1.1);
+    background-color: #fff !important;
 }
   
 .author {
@@ -21,13 +27,14 @@
 .header-panel {
     position: fixed;
     width: 100%;
-    background-color: #9e9e9e;
+    background-color: lightskyblue;
     display: flex;
     flex-direction: row;
     justify-content: space-between;
     gap: 1rem;
     height: 64px;
     padding: 16px;
+    z-index: 999;
 }
 
 .header-panel-inputs {
